+++ src/projectileEmitter.h
@@ -5,7 +5,7 @@ class projectileEmitter : public emitter
 {
 	float frequency = 100.0f;//ms between shots
 	float speed = 5.0f;
-	float lstF = 0.0f;
+	long lstF = 0.0f;
 	bool fire = false;
 	glm::vec2 dir;
 	ofSoundPlayer shootfx;
